#96 addメソッドリファクタリング
